refactor(jans-cedarling): move test mod from init/test.rs into init/policy_store.rs
