+++ jans-cedarling/cedarling/src/init/mod.rs
@@ -16,6 +16,3 @@ pub(crate) mod service_config;
 pub(crate) mod service_factory;
 
 pub(crate) use service_factory::ServiceFactory;
-
-#[cfg(test)]
-mod test;
